chatbot-langgraph/pyproject.toml
[project]
name = "chatbot-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "langsmith>=0.4.29",
    "mcp>=1.14.1",
    "python-dotenv>=1.1.1",
]
